Add hash-based value index
This index is meant for opaque identifiers that are can only be
retrieved via equality lookups. If a string type contains the #id
attribute in a schema, the index factory chooses a new hash-based
implementation.

The hash index is a vector of digests, where each digest is trimmed to a
specific byte length. This works only if the hash function distributes
its value uniformly over the entire output bits. Whether our current
choice of hash function (xxhash64) is an open question that needs to
figured out.